+++ src/main/java/kr/hhplus/be/server/user/domain/userBalance/UserBalanceRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.domain.user;
+package kr.hhplus.be.server.user.domain.userBalance;
 
 
 import java.util.Optional;
