+++ src/main/java/kr/hhplus/be/server/user/infrastructure/persistence/user/SpringDataUserJpaRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.infrastructure.persistence.user;
+package kr.hhplus.be.server.user.infrastructure.persistence.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
